backend-django/wallstash/requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
black==26.1.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.4
defusedxml==0.7.1
Django==5.2.10
django-appconf==1.2.0
django-filter==25.2
django-imagekit==6.0.0
django-taggit==6.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-nested-routers==0.95.0
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==26.0
pathspec==1.0.4
pika==1.3.2
pilkit==3.0
pillow==12.1.0
platformdirs==4.9.2
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyaes==1.6.1
pyasn1==0.6.2
pycparser==3.0
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python3-openid==3.2.0
pytokens==0.4.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
social-auth-app-django==5.7.0
social-auth-core==4.8.3
sqlparse==0.5.5
Telethon==1.42.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
